# openapi.yaml
openapi: 3.0.3
info:
  title: Web屋姫 API
  description: |
    Web屋姫のREST APIドキュメント。
    ユーザー管理とサマリー（概要欄）管理のエンドポイントを提供します。
  version: 1.0.0
  contact:
    name: o-ga09
    url: https://github.com/o-ga09/web-ya-hime

servers:
  - url: http://localhost:8080
    description: ローカル開発環境

tags:
  - name: health
    description: ヘルスチェック
  - name: users
    description: ユーザー管理
  - name: summaries
    description: サマリー（概要欄）管理
  - name: categories
    description: カテゴリ管理
  - name: subcategories
    description: サブカテゴリ管理

paths:
  /health:
    get:
      tags:
        - health
      summary: ヘルスチェック
      description: アプリケーションの稼働状態を確認します
      operationId: healthCheck
      responses:
        '200':
          description: 正常稼働中
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /db-health:
    get:
      tags:
        - health
      summary: データベースヘルスチェック
      description: データベース接続の状態を確認します
      operationId: dbHealthCheck
      responses:
        '200':
          description: データベース接続正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /users:
    post:
      tags:
        - users
      summary: ユーザー作成
      description: 新しいユーザーを作成します
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveUserRequest'
      responses:
        '200':
          description: ユーザー作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                    description: 作成されたユーザーのID
                    example: "123e4567-e89b-12d3-a456-426614174000"
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - users
      summary: ユーザー一覧取得
      description: 登録されているすべてのユーザーを取得します
      operationId: listUsers
      responses:
        '200':
          description: ユーザー一覧取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUserResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /users/{id}:
    put:
      tags:
        - users
      summary: ユーザー更新
      description: 指定されたIDのユーザー情報を更新します
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: ユーザーID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveUserRequest'
      responses:
        '200':
          description: ユーザー更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                    description: 更新されたユーザーのID
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - users
      summary: ユーザー詳細取得
      description: 指定されたIDのユーザー詳細情報を取得します
      operationId: getUserDetail
      parameters:
        - name: id
          in: path
          required: true
          description: ユーザーID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: ユーザー詳細取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailUserResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - users
      summary: ユーザー削除
      description: 指定されたIDのユーザーを削除します
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: ユーザーID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: ユーザー削除成功
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /summaries:
    post:
      tags:
        - summaries
      summary: サマリー作成
      description: 新しいサマリー（概要欄）を作成します
      operationId: createSummary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveSummaryRequest'
      responses:
        '200':
          description: サマリー作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary_id:
                    type: string
                    format: uuid
                    description: 作成されたサマリーのID
                    example: "123e4567-e89b-12d3-a456-426614174001"
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - summaries
      summary: サマリー一覧取得
      description: |
        登録されているサマリーを取得します。
        カテゴリで絞り込み、ページネーションに対応しています。
      operationId: listSummaries
      parameters:
        - name: category
          in: query
          required: false
          description: 'カテゴリで絞り込み（例: 雑談、ゲーム（スプラ）、ゲーム（APEX）など）'
          schema:
            type: string
            example: "雑談"
        - name: category_id
          in: query
          required: false
          description: 'カテゴリIDで絞り込み'
          schema:
            type: string
            format: uuid
        - name: subcategory_id
          in: query
          required: false
          description: |
            サブカテゴリIDで絞り込み。
            subcategory_idのみ指定した場合、親カテゴリIDは自動で補完されます。
            category_idとsubcategory_idを両方指定した場合、組み合わせが正しいかバリデーションされます。
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          description: '取得件数（デフォルト: 20、最大: 100）'
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 50
        - name: offset
          in: query
          required: false
          description: 'オフセット（デフォルト: 0）'
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
      responses:
        '200':
          description: サマリー一覧取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSummaryResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /summaries/{id}:
    put:
      tags:
        - summaries
      summary: サマリー更新
      description: 指定されたIDのサマリーを更新します
      operationId: updateSummary
      parameters:
        - name: id
          in: path
          required: true
          description: サマリーID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveSummaryRequest'
      responses:
        '200':
          description: サマリー更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary_id:
                    type: string
                    format: uuid
                    description: 更新されたサマリーのID
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - summaries
      summary: サマリー詳細取得
      description: 指定されたIDのサマリー詳細情報を取得します
      operationId: getSummaryDetail
      parameters:
        - name: id
          in: path
          required: true
          description: サマリーID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: サマリー詳細取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailSummary'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - summaries
      summary: サマリー削除
      description: 指定されたIDのサマリーを削除します
      operationId: deleteSummary
      parameters:
        - name: id
          in: path
          required: true
          description: サマリーID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: サマリー削除成功
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /categories:
    post:
      tags:
        - categories
      summary: カテゴリ作成
      description: 新しいカテゴリを作成します
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveCategoryRequest'
      responses:
        '200':
          description: カテゴリ作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  category_id:
                    type: string
                    format: uuid
                    description: 作成されたカテゴリのID
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - categories
      summary: カテゴリ一覧取得
      description: 登録されているすべてのカテゴリを取得します
      operationId: listCategories
      responses:
        '200':
          description: カテゴリ一覧取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListCategoryResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /categories/{id}:
    put:
      tags:
        - categories
      summary: カテゴリ更新
      description: 指定されたIDのカテゴリを更新します
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          description: カテゴリID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveCategoryRequest'
      responses:
        '200':
          description: カテゴリ更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  category_id:
                    type: string
                    format: uuid
                    description: 更新されたカテゴリのID
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - categories
      summary: カテゴリ詳細取得
      description: 指定されたIDのカテゴリ詳細情報を取得します
      operationId: getCategoryDetail
      parameters:
        - name: id
          in: path
          required: true
          description: カテゴリID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: カテゴリ詳細取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailCategoryResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - categories
      summary: カテゴリ削除
      description: 指定されたIDのカテゴリを削除します
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          description: カテゴリID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: カテゴリ削除成功
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /subcategories:
    post:
      tags:
        - subcategories
      summary: サブカテゴリ作成
      description: 新しいサブカテゴリを作成します
      operationId: createSubcategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveSubcategoryRequest'
      responses:
        '200':
          description: サブカテゴリ作成成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  subcategory_id:
                    type: string
                    format: uuid
                    description: 作成されたサブカテゴリのID
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      tags:
        - subcategories
      summary: サブカテゴリ一覧取得
      description: |
        登録されているサブカテゴリを取得します。
        カテゴリIDを指定することで、そのカテゴリに属するサブカテゴリのみを取得できます。
      operationId: listSubcategories
      parameters:
        - name: category_id
          in: query
          required: false
          description: カテゴリID（指定すると、そのカテゴリに属するサブカテゴリのみを取得）
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: サブカテゴリ一覧取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSubcategoryResponse'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /subcategories/{id}:
    put:
      tags:
        - subcategories
      summary: サブカテゴリ更新
      description: 指定されたIDのサブカテゴリを更新します
      operationId: updateSubcategory
      parameters:
        - name: id
          in: path
          required: true
          description: サブカテゴリID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveSubcategoryRequest'
      responses:
        '200':
          description: サブカテゴリ更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  subcategory_id:
                    type: string
                    format: uuid
                    description: 更新されたサブカテゴリのID
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - subcategories
      summary: サブカテゴリ詳細取得
      description: 指定されたIDのサブカテゴリ詳細情報を取得します
      operationId: getSubcategoryDetail
      parameters:
        - name: id
          in: path
          required: true
          description: サブカテゴリID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: サブカテゴリ詳細取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailSubcategoryResponse'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - subcategories
      summary: サブカテゴリ削除
      description: 指定されたIDのサブカテゴリを削除します
      operationId: deleteSubcategory
      parameters:
        - name: id
          in: path
          required: true
          description: サブカテゴリID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: サブカテゴリ削除成功
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    SaveUserRequest:
      type: object
      required:
        - name
        - email
        - user_type
      properties:
        name:
          type: string
          maxLength: 100
          description: ユーザー名
          example: "山田太郎"
        email:
          type: string
          format: email
          maxLength: 255
          description: メールアドレス
          example: "yamada@example.com"
        user_type:
          type: string
          description: ユーザータイプ
          example: "admin"

    ListUserResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
          description: ユーザーの総数
          example: 10

    DetailUserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: ユーザーID
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: ユーザー名
          example: "山田太郎"
        email:
          type: string
          format: email
          description: メールアドレス
          example: "yamada@example.com"
        user_type:
          type: string
          description: ユーザータイプ
          example: "admin"

    SaveSummaryRequest:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
          maxLength: 255
          description: タイトル
          example: "プロジェクト概要"
        description:
          type: string
          maxLength: 5000
          description: 説明
          example: "このプロジェクトの概要を記載します"
        content:
          type: string
          description: コンテンツ本文
          example: "詳細な内容がここに入ります。"
        category_id:
          type: string
          format: uuid
          description: カテゴリID
          nullable: true
        subcategory_id:
          type: string
          format: uuid
          description: |
            サブカテゴリID。
            subcategory_idのみ指定した場合、親カテゴリIDは自動で補完されます。
            category_idとsubcategory_idを両方指定した場合、組み合わせが正しいかバリデーションされます。
          nullable: true
        user_id:
          type: string
          format: uuid
          description: 作成者のユーザーID
          example: "123e4567-e89b-12d3-a456-426614174000"

    DetailSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: サマリーID
          example: "123e4567-e89b-12d3-a456-426614174001"
        user:
          $ref: '#/components/schemas/User'
          nullable: true
        title:
          type: string
          description: タイトル
          example: "プロジェクト概要"
        description:
          type: string
          description: 説明
          example: "このプロジェクトの概要を記載します"
        content:
          type: string
          description: コンテンツ本文
          example: "詳細な内容がここに入ります。"
        category_obj:
          $ref: '#/components/schemas/Category'
          description: カテゴリオブジェクト
          nullable: true
        subcategory:
          $ref: '#/components/schemas/Subcategory'
          description: サブカテゴリオブジェクト
          nullable: true
        created_at:
          type: string
          format: date-time
          description: 作成日時
          example: "2025-12-07 10:30:00"
        updated_at:
          type: string
          format: date-time
          description: 更新日時
          example: "2025-12-07 10:30:00"

    ListSummaryResponse:
      type: object
      properties:
        summaries:
          type: array
          items:
            $ref: '#/components/schemas/DetailSummary'
          description: サマリーの配列
        total:
          type: integer
          description: 総件数（フィルタリング条件に一致する全データ件数）
          example: 125
        limit:
          type: integer
          description: 1回のリクエストで取得する最大件数
          example: 50
        offset:
          type: integer
          description: 取得開始位置（0始まり）
          example: 0
        has_next:
          type: boolean
          description: 次のページが存在するかどうか
          example: true

    SaveCategoryRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
          description: カテゴリ名
          example: "ゲーム"

    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: カテゴリID
        name:
          type: string
          description: カテゴリ名
          example: "ゲーム"
        created_at:
          type: string
          format: date-time
          description: 作成日時
        updated_at:
          type: string
          format: date-time
          description: 更新日時

    ListCategoryResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        total:
          type: integer
          description: カテゴリの総数

    DetailCategoryResponse:
      type: object
      properties:
        category:
          $ref: '#/components/schemas/Category'

    SaveSubcategoryRequest:
      type: object
      required:
        - category_id
        - name
      properties:
        category_id:
          type: string
          format: uuid
          description: 親カテゴリID
        name:
          type: string
          maxLength: 100
          description: サブカテゴリ名
          example: "スプラトゥーン"

    Subcategory:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: サブカテゴリID
        category_id:
          type: string
          format: uuid
          description: 親カテゴリID
        category_name:
          type: string
          description: 親カテゴリ名
          example: "ゲーム"
        name:
          type: string
          description: サブカテゴリ名
          example: "スプラトゥーン"
        created_at:
          type: string
          format: date-time
          description: 作成日時
        updated_at:
          type: string
          format: date-time
          description: 更新日時

    ListSubcategoryResponse:
      type: object
      properties:
        subcategories:
          type: array
          items:
            $ref: '#/components/schemas/Subcategory'
        total:
          type: integer
          description: サブカテゴリの総数

    DetailSubcategoryResponse:
      type: object
      properties:
        subcategory:
          $ref: '#/components/schemas/Subcategory'

    Error:
      type: object
      properties:
        message:
          type: string
          description: エラーメッセージ
          example: "Invalid request parameters"